+++ src/Doctrine/DataType/DateTimeDataType.php
@@ -6,6 +6,8 @@ namespace Novuso\Common\Adapter\Doctrine\DataType;
 
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Types\ConversionException;
+use Doctrine\DBAL\Types\Exception\InvalidType;
+use Doctrine\DBAL\Types\Exception\ValueNotConvertible;
 use Doctrine\DBAL\Types\Type;
 use Exception;
 use Novuso\Common\Domain\Value\DateTime\DateTime;
@@ -46,11 +48,7 @@ final class DateTimeDataType extends Type
         }
 
         if (!($value instanceof DateTime)) {
-            throw ConversionException::conversionFailedInvalidType(
-                $value,
-                'string',
-                [DateTime::class]
-            );
+            throw InvalidType::new($value, 'string', [DateTime::class]);
         }
 
         // @codeCoverageIgnoreStart
@@ -96,10 +94,7 @@ final class DateTimeDataType extends Type
                 $second
             );
         } catch (Throwable $e) {
-            throw ConversionException::conversionFailed(
-                $value,
-                static::TYPE_NAME
-            );
+            throw ValueNotConvertible::new($value, DateTime::class, $e->getMessage(), $e);
         }
     }
 
@@ -109,13 +104,5 @@ final class DateTimeDataType extends Type
     public function getName(): string
     {
         return static::TYPE_NAME;
-    }
-
-    /**
-     * Checks if this type requires a SQL comment hint
-     */
-    public function requiresSQLCommentHint(AbstractPlatform $platform): bool
-    {
-        return true;
     }
 }